Проект собран webpack. Реализован модульный js
