+++ index.html
@@ -9,10 +9,9 @@
 	<link rel="stylesheet" href="css/normalize.css" />
 	<link rel="stylesheet" href="css/style.css" />
 	<script src="https://unpkg.com/swiper@7/swiper-bundle.min.js" defer></script>
-	<script src="js/auth.js" defer></script>
-	<script src="js/partners.js" defer></script>
-	<script src="js/slider.js" defer></script>
-	<script src="js/cart.js" defer></script>
+	<script src="./dist/main.js" defer></script>
+	<script src="src/modules/slider.js" defer></script>
+	<script src="src/modules/cart.js" defer></script>
 
 	<link rel="stylesheet" href="https://unpkg.com/swiper@7/swiper-bundle.min.css"/>
 </head>
