--- js/partners.js
@@ -1,58 +1,0 @@
-'use strict'; 
-const cardsRestaurants = document.querySelector('.cards-restaurants');
-//const 
-
-const renderItems = (data) => {
-  data.forEach((item) => {
-    //деструктуризация:
-    const {image, kitchen, name, price, products, stars, time_of_delivery} = item;
-    const a = document.createElement('a');
-    
-    a.setAttribute('href','./restaurant.html');
-    a.classList.add('card');
-    a.classList.add('card-restaurant');
-    //сохр обьект product в data-атрибуте для получения уникальных свойств каждого ресторана:
-    a.dataset.products = products;
-
-    a.innerHTML = `
-      <img src=${image} alt="${name}" class="card-image" />
-      <div class="card-text">
-        <div class="card-heading">
-          <h3 class="card-title">${name}</h3>
-          <span class="card-tag tag">${time_of_delivery} мин</span>
-        </div>
-        <div class="card-info">
-          <div class="rating">${stars}</div>
-          <div class="price">от ${price} руб</div>
-          <div class="category">${kitchen}</div>
-        </div>
-      </div>`
-
-    a.addEventListener('click', (event) => {
-      event.preventDefault();
-
-      if(!localStorage.getItem('user')) {
-        modalAuth.style.display = 'flex';
-        event.preventDefault();
-      } else {
-         //записываем в localStorage данные выбранного ресторана
-        localStorage.setItem('restaurant', JSON.stringify(item));
-        //переходим на страницу выбранного ресторана
-        window.location.href = 'restaurant.html';
-      }
-    })
-
-    cardsRestaurants.append(a);
-  });
-}
-
-
-
-fetch('./db/partners.json')
-  .then((response) => response.json())
-  .then((data) => {
-    renderItems(data);
-  })
-  .catch((error) => {
-    console.log(error);
-  })
