+++ restaurant.html
@@ -8,9 +8,8 @@
 	<link href="https://fonts.googleapis.com/css?family=Roboto:400,700&display=swap&subset=cyrillic" rel="stylesheet" />
 	<link rel="stylesheet" href="css/normalize.css" />
 	<link rel="stylesheet" href="css/style.css" />
-	<script src="js/auth.js" defer></script>
-	<script src="js/menu.js" defer></script>
-	<script src="js/cart.js" defer></script>
+	<script src="./dist/menu.js" defer></script>
+	<script src="src/modules/cart.js" defer></script>
 
 </head>
