+++ src/index.js
@@ -1,0 +1,10 @@
+
+import auth from "./modules/auth";
+import partners from "./modules/partners";
+import cart from "./modules/cart";
+import slider from "./modules/slider";
+
+auth()
+partners()
+cart()
+slider()
