+++ src/menu.js
@@ -1,0 +1,7 @@
+import auth from "./modules/auth";
+import menu from "./modules/menu";
+import cart from "./modules/cart";
+
+auth()
+menu()
+cart()
